+++ src/processes/standalone/mod.rs
@@ -1,2 +1,3 @@
 pub mod sampler;
 pub mod trimmer;
+pub mod xleave;